Workstation/requirement.txt
astunparse==1.6.3
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
numpy==1.24.4
opencv-python==4.8.0.74
orjson==3.9.2
Pillow==10.0.0
requests==2.31.0
setuptools==67.8.0
tk==0.1.0
ujson==5.8.0
urllib3==2.0.4
wheel==0.38.4
google-api-python-client==2.112.0
google-auth-oauthlib==1.2.0
flask==3.0.1
tqdm==4.66.1
pymongo==4.6.1